third_party/WebKit/LayoutTests/svg/markers/marker-orientation-set.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>
test(function() {
    var svgElement = document.createElementNS("http://www.w3.org/2000/svg","svg");
    var markerElement = document.createElementNS("http://www.w3.org/2000/svg","marker");
    var svgAngle = svgElement.createSVGAngle();
    svgAngle.value = 10;
    markerElement.setOrientToAngle(svgAngle);
    assert_equals(markerElement.getAttribute("orient"), "10");
}, "Tests getAttribute('orient') value after setOrientToAngle");

test(function() {
    var markerElement = document.createElementNS("http://www.w3.org/2000/svg","marker");
    markerElement.setOrientToAuto();
    assert_equals(markerElement.getAttribute("orient"), "auto");
}, "Tests getAttribute('orient') value after setOrientToAuto");
</script>
